feature/event manager

+++ src/wall.cpp
@@ -39,10 +39,14 @@ void Wall::Update(const float dt)
 
 void Wall::Render(sf::RenderWindow& window)
 {
-    shape_.setPosition(top_left_.x * BLOCK_SIZE, top_left_.y * BLOCK_SIZE);
+    shape_.setPosition(
+        static_cast<float>(top_left_.x * BLOCK_SIZE),
+        static_cast<float>(top_left_.y * BLOCK_SIZE)
+    );
+
     shape_.setSize(sf::Vector2f(
-        (bottom_right_.x + 1 - top_left_.x) * BLOCK_SIZE,
-        (bottom_right_.y + 1 - top_left_.y) * BLOCK_SIZE
+        static_cast<float>((bottom_right_.x + 1 - top_left_.x) * BLOCK_SIZE),
+        static_cast<float>((bottom_right_.y + 1 - top_left_.y) * BLOCK_SIZE)
     ));
 
     shape_.setFillColor(sf::Color::Magenta);
